[yt] Refactor phy opt code
